+implemented full win audio support-  check qbplay.cs for usage info
