+++ PlayString.csproj
@@ -48,7 +48,7 @@
   <ItemGroup>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
-    <Compile Include="STR2WAV.cs" />
+    <Compile Include="qbplay.cs" />
   </ItemGroup>
   <ItemGroup>
     <None Include="app.config" />
